Update Menu and retrieveInput Functionality
When retrieveInput returned zeros, they'd be mistaken for errors. The
reverse was also true. This fixes that by adding a pointer that is
updated to signify whether input was successfully received or not. This
also swaps the if else tree with switch statements to make the code
cleaner. Finally, this adds input support for the Prime Factorization
feature.