Improve getNextPrime
Updates the getNextPrime method to handle input of any kind, not just
prime numbers. Instead of checking for the next prime after a prime
number, it'll find the next prime after any number. This improves
efficiency by not having to check if the input was prime or not, saving
on computations done.