Improve Error Checking
Updates the current tests and main file to better detect and handle
errors. Fixes things like an infinite loop in the Prime Factorization
feature if a negative number was added. Also adds the Fibonacci feature
to the main menu. Checking for malloc and realloc failures will be added
in a later commit.