Rewrite Input System and Error Handling
Rewrite the input system to handle alphanumeric characters. Now it can
return purely numeric characters or alphanumeric characters. This is in
preparation for the Base Conversion feature. This also updates the error
handling in the main file to remove unnecessary checks. Finally, this
adds the menu for the Base Conversion feature.